+++ src/main/java/finalmission/controller/dto/ReservationResponse.java
@@ -1,4 +1,4 @@
-package finalmission.controller;
+package finalmission.controller.dto;
 
 import finalmission.domain.Reservation;
 import java.time.LocalDate;
